Move exit status code
